add preserve-on-error for docker

+++ aminator/plugins/volume/docker.py
@@ -45,6 +45,8 @@ class DockerVolumePlugin(BaseVolumePlugin):
 
     def __exit__(self, exc_type, exc_value, trace):
         if exc_type: log.exception("Exception: {0}: {1}".format(exc_type.__name__,exc_value))
+        if exc_type and self._config.context.get("preserve_on_error", False):
+            return False
         self._cloud.detach_volume(self._blockdevice)
         self._cloud.delete_volume()
         return False
